code/requirements.txt
newsapi-python
json2html